Check for special values

biginteger and bigfloat support missing values (via NA_biginteger_ and NA_bigfloat_ respectively).

bigfloat additionally supports positive and negative infinity and 'Not a Number' values. Usually these are the result of a calculation, but they can also be created manually by casting from numeric to bigfloat.

These functions check for the presence of these special values. The base R documentation can be found at is.na() and is.finite().

Value

A logical vector.

See also

Examples

x <- bigfloat(c(0, NA, Inf, -Inf, NaN))

is.na(x)

#> [1] FALSE  TRUE FALSE FALSE  TRUE

is.finite(x)

#> [1]  TRUE FALSE FALSE FALSE FALSE

is.infinite(x)

#> [1] FALSE FALSE  TRUE  TRUE FALSE

is.nan(x)

#> [1] FALSE FALSE FALSE FALSE  TRUE